type changes to smaller sizes for pixel info

+++ source/ROOTFileManager.h
@@ -55,12 +55,12 @@ namespace Qpix {
 
             std::vector< int > pixel_x_;
             std::vector< int > pixel_y_;
-            std::vector< std::vector < double > > pixel_reset_;
-            std::vector< std::vector < double > > pixel_tslr_;
+            std::vector< std::vector < float > > pixel_reset_;
+            std::vector< std::vector < float > > pixel_tslr_;
             // std::vector< std::vector< std::vector < int > > > pixel_reset_truth_track_id_;
             // std::vector< std::vector< std::vector < double > > > pixel_reset_truth_weight_;
-            std::vector< std::vector< std::vector < int > > > pixel_reset_truth_track_id_;
-            std::vector< std::vector< std::vector < int > > > pixel_reset_truth_weight_;
+            std::vector< std::vector< std::vector < u_int16_t > > > pixel_reset_truth_track_id_;
+            std::vector< std::vector< std::vector < short > > > pixel_reset_truth_weight_;
 
             TBranch * tbranch_x_;
             TBranch * tbranch_y_;
